Reorganizing the Files in IO
